refactor: split metrics module for better cohesion
- Extracted core metric types (FScore, Precision, Recall, Beta, ConfidenceLevel) to core_metrics module
- Moved FScoreDataPoint to its own data_point module
- Extracted MetricsError to separate errors module
- Improved module organization with clear separation of concerns
- All tests passing after refactoring
